requirements.txt
tensorflow
tensorflow_datasets
flask
flask-restful
keras
matplotlib
pandas
pillow
numpy
requests
